Web App/requirements.txt
# Web framework
streamlit

# System / utility
requests
uuid

# OpenAI SDK
openai

# Databricks Vector Search SDK
databricks-vectorsearch

# Optional (recommended)
databricks-sdk
